requirements.txt
absl-py==0.7.1
alabaster==0.7.12
asn1crypto==0.24.0
astor==0.6.2
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
bleach==1.5.0
bokeh==0.12.15
certifi==2019.3.9
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.7.0
cryptography==2.3
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==1.1.3
decorator==4.4.0
defusedxml==0.5.0
distributed==1.21.6
docutils==0.14
entrypoints==0.3
Flask==1.0.2
gast==0.2.0
gevent==1.4.0
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.11.0
h5py==2.7.1
heapdict==1.0.0
html5lib==0.9999999
idna==2.6
imageio==2.5.0
imagesize==1.1.0
imgaug==0.2.8
ipykernel==5.1.0
ipyparallel==6.2.3
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.1.6
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
locket==0.2.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
memory-profiler==0.52.0
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
mock==2.0.0
msgpack==0.6.1
msgpack-python==0.5.6
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nose==1.3.7
notebook==5.7.8
numpy==1.16.2
olefile==0.45.1
opencv-python==4.0.0.21
packaging==19.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.8
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
protobuf==3.7.0
psutil==5.4.5
ptyprocess==0.6.0
pycparser==2.18
pydensecrf==1.0rc2
pydot==1.2.4
Pygments==2.3.1
pyOpenSSL==17.5.0
pyparsing==2.3.1
pypng==0.0.19
pyrsistent==0.14.11
PySocks==1.6.7
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.20.0
# Editable install with no version control (scikit-image==0.15.0)
-e /home/students/student2_1/.local/lib/python3.6/site-packages
scikit-learn==0.19.1
scipy==1.2.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
snowballstemmer==1.2.1
sortedcontainers==1.5.9
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
tblib==1.3.2
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==6.0.1
tqdm==4.23.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
zict==0.1.3
